AHHA_240830.01

+++ AHHA.Domain/Entities/Admin/AdmUser.cs
@@ -6,7 +6,7 @@ namespace AHHA.Core.Entities.Admin
     public class AdmUser
     {
         [Key]
-        public Int32 UserId { get; set; }
+        public Int16 UserId { get; set; }
 
         public string UserCode { get; set; }
         public string UserName { get; set; }
@@ -27,7 +27,7 @@ namespace AHHA.Core.Entities.Admin
         [NotMapped]
         public DateTime? EditDate { get; set; }
 
-        public Int32 UserGroupId { get; set; }
+        public Int16 UserGroupId { get; set; }
         public string RefreshToken { get; set; }
         public DateTime? RefreshTokenExpiry { get; set; }
     }
