+++ AHHA.Domain/Entities/Admin/AdmUserGroup.cs
@@ -6,7 +6,7 @@ namespace AHHA.Core.Entities.Admin
     public class AdmUserGroup
     {
         [Key]
-        public Int32 UserGroupId { get; set; }
+        public Int16 UserGroupId { get; set; }
 
         public string UserGroupCode { get; set; }
         public string UserGroupName { get; set; }
