+++ AHHA.Domain/Models/Admin/UserGroupViewModel.cs
@@ -2,7 +2,7 @@
 {
     public class UserGroupViewModel
     {
-        public Int32 UserGroupId { get; set; }
+        public Int16 UserGroupId { get; set; }
 
         public string UserGroupCode { get; set; }
         public string UserGroupName { get; set; }
