+++ AHHA.Domain/Models/Admin/UserViewModel.cs
@@ -2,8 +2,7 @@
 {
     public class UserViewModel
     {
-        public Int32 UserId { get; set; }
-
+        public Int16 UserId { get; set; }
         public string UserCode { get; set; }
         public string UserName { get; set; }
         public string UserPassword { get; set; }
@@ -14,7 +13,7 @@
         public DateTime CreateDate { get; set; }
         public Int32? EditById { get; set; }
         public DateTime? EditDate { get; set; }
-        public Int32 UserGroupId { get; set; }
+        public Int16 UserGroupId { get; set; }
         public string UserGroupCode { get; set; }
         public string UserGroupName { get; set; }
         public string CreateBy { get; set; }
